Add ReLU in the kernel. Combine with the convolution

+++ hls_target.h
@@ -12,9 +12,9 @@
 
 
 void hls_target(
-		uint16_t *arg_0,//[32*124*32],
-		uint8_t *arg_1,//[34*126*32],
-		uint8_t *arg_2,//[32*32*9]
+		uint32_t *arg_0,//[32*124*32],
+		uint16_t *arg_1,//[34*126*32],
+		int16_t *arg_2,//[32*32*9]
 		uint8_t Ksz,
 		uint8_t X_n,
 		uint8_t Y_n,
